# Conflicts:
#	consensus/merge/src/main/java/org/hyperledger/besu/consensus/merge/PostMergeContext.java
#	consensus/merge/src/main/java/org/hyperledger/besu/consensus/merge/blockcreation/MergeCoordinator.java
#	ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/AbstractEngineForkchoiceUpdated.java
#	ethereum/core/src/main/java/org/hyperledger/besu/ethereum/bonsai/trielog/AbstractTrieLogManager.java

+++ acceptance-tests/dsl/src/main/java/org/hyperledger/besu/tests/acceptance/dsl/node/ThreadBesuNodeRunner.java
@@ -28,7 +28,9 @@ import org.hyperledger.besu.crypto.KeyPairUtil;
 import org.hyperledger.besu.cryptoservices.KeyPairSecurityModule;
 import org.hyperledger.besu.cryptoservices.NodeKey;
 import org.hyperledger.besu.ethereum.GasLimitCalculator;
+import org.hyperledger.besu.ethereum.api.ApiConfiguration;
 import org.hyperledger.besu.ethereum.api.graphql.GraphQLConfiguration;
+import org.hyperledger.besu.ethereum.core.ImmutableMiningParameters;
 import org.hyperledger.besu.ethereum.eth.EthProtocolConfiguration;
 import org.hyperledger.besu.ethereum.eth.sync.SynchronizerConfiguration;
 import org.hyperledger.besu.ethereum.eth.transactions.ImmutableTransactionPoolConfiguration;
@@ -36,24 +38,38 @@ import org.hyperledger.besu.ethereum.eth.transactions.TransactionPoolConfigurati
 import org.hyperledger.besu.ethereum.p2p.peers.EnodeURLImpl;
 import org.hyperledger.besu.ethereum.storage.keyvalue.KeyValueStorageProvider;
 import org.hyperledger.besu.ethereum.storage.keyvalue.KeyValueStorageProviderBuilder;
+import org.hyperledger.besu.ethereum.transaction.TransactionSimulator;
+import org.hyperledger.besu.ethereum.worldstate.DataStorageConfiguration;
 import org.hyperledger.besu.evm.internal.EvmConfiguration;
 import org.hyperledger.besu.metrics.MetricsSystemFactory;
 import org.hyperledger.besu.metrics.ObservableMetricsSystem;
 import org.hyperledger.besu.plugin.data.EnodeURL;
 import org.hyperledger.besu.plugin.services.BesuConfiguration;
 import org.hyperledger.besu.plugin.services.BesuEvents;
+import org.hyperledger.besu.plugin.services.BlockchainService;
+import org.hyperledger.besu.plugin.services.PermissioningService;
 import org.hyperledger.besu.plugin.services.PicoCLIOptions;
+import org.hyperledger.besu.plugin.services.PrivacyPluginService;
+import org.hyperledger.besu.plugin.services.RpcEndpointService;
 import org.hyperledger.besu.plugin.services.SecurityModuleService;
 import org.hyperledger.besu.plugin.services.StorageService;
+import org.hyperledger.besu.plugin.services.TransactionPoolValidatorService;
+import org.hyperledger.besu.plugin.services.TransactionSelectionService;
+import org.hyperledger.besu.plugin.services.TransactionSimulationService;
 import org.hyperledger.besu.plugin.services.storage.rocksdb.RocksDBPlugin;
 import org.hyperledger.besu.services.BesuConfigurationImpl;
 import org.hyperledger.besu.services.BesuEventsImpl;
 import org.hyperledger.besu.services.BesuPluginContextImpl;
+import org.hyperledger.besu.services.BlockchainServiceImpl;
 import org.hyperledger.besu.services.PermissioningServiceImpl;
 import org.hyperledger.besu.services.PicoCLIOptionsImpl;
+import org.hyperledger.besu.services.PrivacyPluginServiceImpl;
 import org.hyperledger.besu.services.RpcEndpointServiceImpl;
 import org.hyperledger.besu.services.SecurityModuleServiceImpl;
 import org.hyperledger.besu.services.StorageServiceImpl;
+import org.hyperledger.besu.services.TransactionPoolValidatorServiceImpl;
+import org.hyperledger.besu.services.TransactionSelectionServiceImpl;
+import org.hyperledger.besu.services.TransactionSimulationServiceImpl;
 
 import java.io.File;
 import java.nio.file.Path;
@@ -85,29 +101,51 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
       final BesuNode node,
       final StorageServiceImpl storageService,
       final SecurityModuleServiceImpl securityModuleService,
-      final BesuConfiguration commonPluginConfiguration) {
+      final TransactionSimulationServiceImpl transactionSimulationServiceImpl,
+      final TransactionSelectionServiceImpl transactionSelectionServiceImpl,
+      final TransactionPoolValidatorServiceImpl transactionPoolValidatorServiceImpl,
+      final BlockchainServiceImpl blockchainServiceImpl,
+      final RpcEndpointServiceImpl rpcEndpointServiceImpl,
+      final BesuConfiguration commonPluginConfiguration,
+      final PermissioningServiceImpl permissioningService) {
     final CommandLine commandLine = new CommandLine(CommandSpec.create());
     final BesuPluginContextImpl besuPluginContext = new BesuPluginContextImpl();
     besuPluginContext.addService(StorageService.class, storageService);
     besuPluginContext.addService(SecurityModuleService.class, securityModuleService);
     besuPluginContext.addService(PicoCLIOptions.class, new PicoCLIOptionsImpl(commandLine));
+    besuPluginContext.addService(RpcEndpointService.class, rpcEndpointServiceImpl);
+    besuPluginContext.addService(
+        TransactionSelectionService.class, transactionSelectionServiceImpl);
+    besuPluginContext.addService(
+        TransactionPoolValidatorService.class, transactionPoolValidatorServiceImpl);
+    besuPluginContext.addService(
+        TransactionSimulationService.class, transactionSimulationServiceImpl);
+    besuPluginContext.addService(BlockchainService.class, blockchainServiceImpl);
+    besuPluginContext.addService(BesuConfiguration.class, commonPluginConfiguration);
 
-    final Path pluginsPath = node.homeDirectory().resolve("plugins");
-    final File pluginsDirFile = pluginsPath.toFile();
-    if (!pluginsDirFile.isDirectory()) {
-      pluginsDirFile.mkdirs();
-      pluginsDirFile.deleteOnExit();
+    final Path pluginsPath;
+    final String pluginDir = System.getProperty("besu.plugins.dir");
+    if (pluginDir == null || pluginDir.isEmpty()) {
+      pluginsPath = node.homeDirectory().resolve("plugins");
+      final File pluginsDirFile = pluginsPath.toFile();
+      if (!pluginsDirFile.isDirectory()) {
+        pluginsDirFile.mkdirs();
+        pluginsDirFile.deleteOnExit();
+      }
+      System.setProperty("besu.plugins.dir", pluginsPath.toString());
+    } else {
+      pluginsPath = Path.of(pluginDir);
     }
-    System.setProperty("besu.plugins.dir", pluginsPath.toString());
-    besuPluginContext.registerPlugins(pluginsPath);
-
-    commandLine.parseArgs(node.getConfiguration().getExtraCLIOptions().toArray(new String[0]));
 
     besuPluginContext.addService(BesuConfiguration.class, commonPluginConfiguration);
+    besuPluginContext.addService(PermissioningService.class, permissioningService);
+    besuPluginContext.addService(PrivacyPluginService.class, new PrivacyPluginServiceImpl());
+
+    besuPluginContext.registerPlugins(pluginsPath);
+    commandLine.parseArgs(node.getConfiguration().getExtraCLIOptions().toArray(new String[0]));
 
     // register built-in plugins
     new RocksDBPlugin().register(besuPluginContext);
-
     return besuPluginContext;
   }
 
@@ -125,15 +163,43 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
 
     final StorageServiceImpl storageService = new StorageServiceImpl();
     final SecurityModuleServiceImpl securityModuleService = new SecurityModuleServiceImpl();
+    final TransactionSimulationServiceImpl transactionSimulationServiceImpl =
+        new TransactionSimulationServiceImpl();
+    final TransactionSelectionServiceImpl transactionSelectionServiceImpl =
+        new TransactionSelectionServiceImpl();
+    final TransactionPoolValidatorServiceImpl transactionPoolValidatorServiceImpl =
+        new TransactionPoolValidatorServiceImpl();
+    final BlockchainServiceImpl blockchainServiceImpl = new BlockchainServiceImpl();
+    final RpcEndpointServiceImpl rpcEndpointServiceImpl = new RpcEndpointServiceImpl();
     final Path dataDir = node.homeDirectory();
-    final BesuConfiguration commonPluginConfiguration =
-        new BesuConfigurationImpl(dataDir, dataDir.resolve(DATABASE_PATH));
+    final BesuConfigurationImpl commonPluginConfiguration = new BesuConfigurationImpl();
+    final PermissioningServiceImpl permissioningService = new PermissioningServiceImpl();
+
+    final var miningParameters =
+        ImmutableMiningParameters.builder()
+            .from(node.getMiningParameters())
+            .transactionSelectionService(transactionSelectionServiceImpl)
+            .build();
+    commonPluginConfiguration.init(
+        dataDir,
+        dataDir.resolve(DATABASE_PATH),
+        node.getDataStorageConfiguration(),
+        miningParameters);
     final BesuPluginContextImpl besuPluginContext =
         besuPluginContextMap.computeIfAbsent(
             node,
             n ->
                 buildPluginContext(
-                    node, storageService, securityModuleService, commonPluginConfiguration));
+                    node,
+                    storageService,
+                    securityModuleService,
+                    transactionSimulationServiceImpl,
+                    transactionSelectionServiceImpl,
+                    transactionPoolValidatorServiceImpl,
+                    blockchainServiceImpl,
+                    rpcEndpointServiceImpl,
+                    commonPluginConfiguration,
+                    permissioningService));
 
     GlobalOpenTelemetry.resetForTest();
     final ObservableMetricsSystem metricsSystem =
@@ -164,7 +230,9 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
 
     final TransactionPoolConfiguration txPoolConfig =
         ImmutableTransactionPoolConfiguration.builder()
+            .from(node.getTransactionPoolConfiguration())
             .strictTransactionReplayProtectionEnabled(node.isStrictTxReplayProtectionEnabled())
+            .transactionPoolValidatorService(transactionPoolValidatorServiceImpl)
             .build();
 
     final int maxPeers = 25;
@@ -172,11 +240,12 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
     builder
         .synchronizerConfiguration(new SynchronizerConfiguration.Builder().build())
         .dataDirectory(node.homeDirectory())
-        .miningParameters(node.getMiningParameters())
+        .miningParameters(miningParameters)
         .privacyParameters(node.getPrivacyParameters())
         .nodeKey(new NodeKey(new KeyPairSecurityModule(KeyPairUtil.loadKeyPair(dataDir))))
         .metricsSystem(metricsSystem)
         .transactionPoolConfiguration(txPoolConfig)
+        .dataStorageConfiguration(DataStorageConfiguration.DEFAULT_FOREST_CONFIG)
         .ethProtocolConfiguration(EthProtocolConfiguration.defaultConfig())
         .clock(Clock.systemUTC())
         .isRevertReasonEnabled(node.isRevertReasonEnabled())
@@ -187,7 +256,6 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
                 .map(pkiConfig -> new PkiBlockCreationConfigurationProvider().load(pkiConfig)))
         .evmConfiguration(EvmConfiguration.DEFAULT)
         .maxPeers(maxPeers)
-        .lowerBoundPeers(maxPeers)
         .maxRemotelyInitiatedPeers(15)
         .networkConfiguration(node.getNetworkingConfiguration())
         .randomPeerPriority(false);
@@ -198,8 +266,13 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
 
     final BesuController besuController = builder.build();
 
+    initTransactionSimulationService(
+        transactionSimulationServiceImpl, besuController, node.getApiConfiguration());
+    initBlockchainService(blockchainServiceImpl, besuController);
+
     final RunnerBuilder runnerBuilder = new RunnerBuilder();
     runnerBuilder.permissioningConfiguration(node.getPermissioningConfiguration());
+    runnerBuilder.apiConfiguration(node.getApiConfiguration());
 
     runnerBuilder
         .vertx(Vertx.vertx())
@@ -214,7 +287,7 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
         .jsonRpcIpcConfiguration(node.jsonRpcIpcConfiguration())
         .dataDir(node.homeDirectory())
         .metricsSystem(metricsSystem)
-        .permissioningService(new PermissioningServiceImpl())
+        .permissioningService(permissioningService)
         .metricsConfiguration(node.getMetricsConfiguration())
         .p2pEnabled(node.isP2pEnabled())
         .p2pTLSConfiguration(node.getTLSConfiguration())
@@ -223,15 +296,14 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
             node.getStaticNodes().stream()
                 .map(EnodeURLImpl::fromString)
                 .collect(Collectors.toList()))
-        .besuPluginContext(new BesuPluginContextImpl())
+        .besuPluginContext(besuPluginContext)
         .autoLogBloomCaching(false)
         .storageProvider(storageProvider)
-        .rpcEndpointService(new RpcEndpointServiceImpl());
+        .rpcEndpointService(rpcEndpointServiceImpl);
     node.engineRpcConfiguration().ifPresent(runnerBuilder::engineJsonRpcConfiguration);
 
-    final Runner runner = runnerBuilder.build();
-
     besuPluginContext.beforeExternalServices();
+    final Runner runner = runnerBuilder.build();
 
     runner.startExternalServices();
 
@@ -248,6 +320,25 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
 
     besuRunners.put(node.getName(), runner);
     MDC.remove("node");
+  }
+
+  private void initBlockchainService(
+      final BlockchainServiceImpl blockchainServiceImpl, final BesuController besuController) {
+    blockchainServiceImpl.init(
+        besuController.getProtocolContext(), besuController.getProtocolSchedule());
+  }
+
+  private void initTransactionSimulationService(
+      final TransactionSimulationServiceImpl transactionSimulationService,
+      final BesuController besuController,
+      final ApiConfiguration apiConfiguration) {
+    transactionSimulationService.init(
+        besuController.getProtocolContext().getBlockchain(),
+        new TransactionSimulator(
+            besuController.getProtocolContext().getBlockchain(),
+            besuController.getProtocolContext().getWorldStateArchive(),
+            besuController.getProtocolSchedule(),
+            apiConfiguration.getGasCap()));
   }
 
   @Override
